refactor(format): mv CommonSettings to the root of the crate to reuse it in decoder too.
