+++ src/query/formats/src/field_encoder/json.rs
@@ -19,9 +19,9 @@ use common_io::consts::NULL_BYTES_LOWER;
 use common_io::consts::TRUE_BYTES_LOWER;
 
 use crate::field_encoder::helpers::write_json_string;
-use crate::field_encoder::CommonSettings;
 use crate::field_encoder::FieldEncoderRowBased;
 use crate::field_encoder::FieldEncoderValues;
+use crate::CommonSettings;
 use crate::FileFormatOptionsExt;
 
 pub struct FieldEncoderJSON {
