+++ src/query/formats/src/field_encoder/mod.rs
@@ -20,7 +20,6 @@ mod tsv;
 mod values;
 
 pub use csv::FieldEncoderCSV;
-pub use helpers::CommonSettings;
 pub use json::FieldEncoderJSON;
 pub use row_based::FieldEncoderRowBased;
 pub use tsv::FieldEncoderTSV;
